+++ WheelWizard/Models/RRInfo/RrPlayer.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using WheelWizard.Models.Enums;
 using WheelWizard.Models.MiiImages;
+using WheelWizard.Services;
 using WheelWizard.Utilities;
 
 namespace WheelWizard.Models.RRInfo;
@@ -32,5 +33,5 @@ public class RrPlayer
     }
     
     public bool HasBadges => false;
-    public PlayerWinPosition WinPosition => FriendCodeManager.GetWinPosition(Fc);
+    public BadgeVariant[] BadgeVariants => BadgeManager.Instance.GetBadgeVariants(Fc);
 }
